chore(ast): move `Factor` into seperate file
